+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v2.13.5
 
 * more cache classes are now generic
-
+* added `MetrifiedThreadPoolExecutor.get_queue_length`
+* `ExclusiveWritebackCache.sync` is no longer best effort